% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_PSF.R
\name{impute_PSF}
\alias{impute_PSF}
\title{Function to find missing values in time series data and to impute these missing values}
\usage{
impute_PSF(dataIn)
}
\arguments{
\item{dataIn}{as inpute time series with missing values}
}
\value{
returns the time series with imputed values
}
\description{
Function to find missing values in time series data and to impute these missing values
}
\examples{
# d <- c(1:5,1:5,1:5,1:5,NA,NA,NA,NA,NA,1:5,1:5,1:5,1:5,NA,NA,NA,NA,NA,1:5,1:5,1:5,1:5)
# impute_PSF(d)
}

