% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_patch.R
\name{insert_patch}
\alias{insert_patch}
\title{Function to insert data in time series data}
\usage{
insert_patch(dataIn, pos, dataInsert)
}
\arguments{
\item{dataIn}{as input time series data}

\item{pos}{as position at with insert new data patch to be insert}

\item{dataInsert}{as data to be inserted in time series data "dataIn"}
}
\value{
returns the time series data with inserted data patch
}
\description{
Function to insert data in time series data
}

