% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsf.R
\name{lpsf}
\alias{lpsf}
\title{Function to restrict the legth of dataset in multiples of 24}
\usage{
lpsf(dataIn, n.ahead)
}
\arguments{
\item{dataIn}{as inpute time series data with missing values}

\item{n.ahead}{as xxx}
}
\value{
returns the time series with multiples of 24
}
\description{
Function to restrict the legth of dataset in multiples of 24
}

