% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_patch.R
\name{missing_patch}
\alias{missing_patch}
\title{To find out the missing data in given dataset}
\usage{
missing_patch(dataIn)
}
\arguments{
\item{dataIn}{as input time series data with missing values (NAs)}
}
\value{
patch representing table with missing value patches
}
\description{
To find out the missing data in given dataset
}

